Add account menu
Related to: https://github.com/openmediavault/openmediavault/issues/2200
