+++ deb/openmediavault/usr/share/openmediavault/workbench/navigation.d/account.password.yaml
@@ -1,0 +1,8 @@
+version: "1.0"
+type: navigation-item
+data:
+  path: "account.password"
+  position: 20
+  text: _("Password")
+  icon: password
+  url: "/account/password"
