+++ deb/openmediavault/workbench/src/app/pages/usermgmt/usermgmt.module.ts
@@ -10,8 +10,6 @@ import { UsermgmtRoutingModule } from '~/app/pages/usermgmt/usermgmt-routing.mod
 import { UserDatatablePageComponent } from '~/app/pages/usermgmt/users/user-datatable-page.component';
 import { UserFormPageComponent } from '~/app/pages/usermgmt/users/user-form-page.component';
 import { UserImportFormPageComponent } from '~/app/pages/usermgmt/users/user-import-form-page.component';
-import { UserPasswordFormPageComponent } from '~/app/pages/usermgmt/users/user-password-form-page.component';
-import { UserProfileFormPageComponent } from '~/app/pages/usermgmt/users/user-profile-form-page.component';
 import { UserSettingsFormPageComponent } from '~/app/pages/usermgmt/users/user-settings-form-page.component';
 import { UserSharedFolderPermissionsDatatablePageComponent } from '~/app/pages/usermgmt/users/user-shared-folder-permissions-datatable-page.component';
 
@@ -19,15 +17,13 @@ import { UserSharedFolderPermissionsDatatablePageComponent } from '~/app/pages/u
   declarations: [
     UserDatatablePageComponent,
     UserFormPageComponent,
-    UserPasswordFormPageComponent,
     UserSettingsFormPageComponent,
     UserImportFormPageComponent,
     GroupFormPageComponent,
     GroupImportFormPageComponent,
     GroupDatatablePageComponent,
     GroupSharedFolderPermissionsDatatablePageComponent,
-    UserSharedFolderPermissionsDatatablePageComponent,
-    UserProfileFormPageComponent
+    UserSharedFolderPermissionsDatatablePageComponent
   ],
   imports: [CommonModule, CoreModule, UsermgmtRoutingModule]
 })
